+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 * minor UI improvement in settings panel: add icons to _Enable all... / Disable all..._ combobox items.
 * minor UI improvement in Model dialog: hide tester's input label when type is set to IDE icon.
 * rework JUnit5 icon (PNG icon to SVG), and add an alternative icon.
-* rework LibreOffice icons (PNG icons to SVG). LibreOffice 7.5 [new icons](https://wiki.documentfoundation.org/ReleaseNotes/7.5#Design) will come later, if I find their source files (help needed).
+* rework LibreOffice icons (PNG icons to SVG). Also added LibreOffice 7.5 [new icons](https://wiki.documentfoundation.org/ReleaseNotes/7.5#Design).
+* add alternative icons for MS Office Word, Excel and PowerPoint files.
+* add an alternative icon for CSV files.
+* add an alternative icon for KeePass files.
 
 ## 2023.1 (2023/01/21)
 * settings: rework the icon which indicates if an IDE restart is needed.